Add native leases support in integration tests
