+++ lease-kubernetes-int-test/kubernetes/job-native.yml
@@ -1,0 +1,18 @@
+apiVersion: batch/v1
+kind: Job
+metadata:
+  name: lease-test
+spec:
+  parallelism: 1
+  completions: 1
+  backoffLimit: 0
+  template:
+    spec:
+      containers:
+        - name: lease-test
+          imagePullPolicy: Never
+          image: pekko-lease-kubernetes-int-test:latest
+          env:
+            - name: LEASE_CLASS
+              value: "org.apache.pekko.coordination.lease.kubernetes.NativeKubernetesLease"
+      restartPolicy: Never
