+++ lease-kubernetes-int-test/src/main/resources/application.conf
@@ -14,4 +14,6 @@ pekko {
       }
     }
   }
+  # Possibility to modify lease class via environment, if not set will take default value
+  coordination.lease.kubernetes.lease-class = ${?LEASE_CLASS}
 }
